Refactored EventBridge S3 events to a new event type. Extracted handler, similar to kinesis.
